backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-multipart==0.0.7
pillow==10.2.0
openai==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic==2.6.0
